refactor: extract merge_queue helpers from conftest and fix smoke_recipe fixture scope
C7-2: Move _make_watcher and _queue_state from tests/execution/conftest.py
to a new tests/execution/_merge_queue_helpers.py module, removing the
top-level import of DefaultMergeQueueWatcher and PRFetchState from
autoskillit.execution.merge_queue.

C7-3: Change smoke_recipe fixture in tests/server/test_smoke_pipeline.py
from scope="module" to function scope. Recipe is a mutable dataclass and
module scope allows test pollution.

Also adds two structural guard tests:
- test_execution_conftest_import_guard: AST check that conftest.py does not
  import merge_queue
- test_smoke_recipe_scope_guard: AST check that smoke_recipe does not use
  scope="module"
