+++ tests/execution/test_merge_queue_ejection.py
@@ -11,7 +11,7 @@ import pytest
 
 import autoskillit.execution.merge_queue as _mq
 from autoskillit.core.types import PRState
-from tests.execution.conftest import _make_watcher, _queue_state
+from tests.execution._merge_queue_helpers import _make_watcher, _queue_state
 
 pytestmark = [pytest.mark.layer("execution"), pytest.mark.small]
